Stop excluding Any from completion type info
Summary:
I was surprised in the previous diff that `(Field) p` printed without a type. It turns out that we explicitly hide `Any` from completion type info, but this doesn't seem necessary.

I blamed to D80627802, but this seems to be an unnecessary change in that diff as well. Removing the exclusion only affected two tests, and the change seems reasonable in both cases.

Reviewed By: grievejia

Differential Revision: D85834600

fbshipit-source-id: 352c8853a1aece01e2761080475af8dcdc9a0dd1