Fix crash on unexpected Var::ZERO during autocomplete (redux)
Summary:
In D85281201 I proposed a targeted fix to deal with Var::ZERO while computing completions. Unfortunately, this fix was incomplete. [See appendix for optional details on the incompleteness]

This diff takes a different approach, which avoids ever dealing with Var::ZERO. Instead of using "for_display" at the answers boundary, we use "deep_force" which also replaces any unsolved Var with Unknown.

We similarly use deep_force when accessing solutions across modules, and performing the same transformation when accessing solutions from "clients" like the LSP feels analogous.

Fixes https://github.com/facebook/pyrefly/issues/1257

Appendix:

I addressed a case where Var::ZERO appeared in expand_var, when the completion item is a method. But there was another case where we see the Var::ZERO in completions: when the completion item is a property, and we call the property to compute the getter's return type.

Reviewed By: grievejia

Differential Revision: D85834036

fbshipit-source-id: ae78a9357a527a0280e4900355209b06cecfbf7b