Add a new test case for unifying quantified with loop recursive
Summary:
We don't actually want to do this. It is wrong because we can wind up pinning
the loop recursive var to a type completely inconsistent with the prior type
(that's what happens in the error message).

On top of that, it creates a scenario where unification could result in a
`LoopRecursive` behaving like a `Quantified` which is not desirable because I
want to make `LoopRecursive` vars behave pretty differently, almost more
as if they just "were" the prior type.

Reviewed By: rchen152

Differential Revision: D85785748

fbshipit-source-id: 843a67b7f84dcd3a70937542023d3774899f9633