Add is_python_std_lib_file function.
Summary:
See D85781296 for a more detailed summary of the plan for this stack.

In the previous commit, we updated the interpreter query to also include any std lib paths from the interpreter. Now we create the is_python_std_lib_file function to determine if a file path belongs to a stdlib installation.

One caveat here is that we still have to do some string matching here. I was attempting to use Path::starts_with to componentize the paths and compare that way. However, I found that many times paths for Windows ended up being escaped and having double backslashes in the path. This led to a collection of components where the path was not actually separated, and the entire path was used as the first component.

In the next diff, we will use this implementation of is_python_std_lib_file in order to properly filter out diagnostic errors form the IDE.

Reviewed By: connernilsen

Differential Revision: D85786299

fbshipit-source-id: 150d47ae5787b5dfd2f401ff37487c26d80482e0