Fix issue with appended type hints in sandbox.
Summary:
Currently there is an issue in the sandbox on the Pyrefly website which affects how type hints are displayed. If you reset the sandbox, the existing type hint will remain, and a new type hint will also be appended. You can see an example of this in the video below.

{F1983133356}

The main issue here had to do with the way that we were updating Monaco. Calling setValue did not actually clear our the state of the hints unless an actual change was detected between the previous content and the new content. Now what we do is use a no-op edit operation in order to force an update.

We also force the wasm state to reset everytime the reset button is called by calling updateAllFiles.

Now type hints will properly clear and be reset rather than being appended to the existing type hint.

{F1983133459}

Reviewed By: kinto0

Differential Revision: D85867468

fbshipit-source-id: f7a4dba372964d03e4d7c6cfd9ff3ddddcab85ac