Fix hover coloring issue in VSCode (#1455)
Summary:
Close https://github.com/facebook/pyrefly/issues/1423

It turns out that we have to add ellipses to the end of the function/method signature. Now the hover highlighting works as expected, except for the coloring on `->` that warrants another look in the future...

<img width="312" height="280" alt="Screenshot 2025-11-02 at 3 04 06 PM" src="https://github.com/user-attachments/assets/688037c3-3532-4dbc-9e39-7fe1be4f7811" />

I also realized that functions and methods were not resolved properly, so I added a patch for that as well.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1455

Reviewed By: rchen152

Differential Revision: D86110486

Pulled By: stroxler

fbshipit-source-id: e5940cc83d34c53efe707644ca714855c47fb7e5