Temporary: panic if a LoopRecursive appears as `want`
Summary:
A LoopRecursive `Var` always represents the type of a *term*, because the
only way to get one is to be looking at a name in scope at the top of a
loop. As a result, it should be impossible for it to appear in an upper
bound position.

This diff is intended to be landed alongside a follow-up diff that makes
this case an internal type error rather than a panic. The rationale is that
the invariant that LoopRecursive can't appear in a `want` position is quite
hard to verify absolutely. This diff (and the CI runs) will verify that
empirically it seems to hold, but using a type error ensures that if there
are edge cases or if we introduce a regression we'll fail gracefully rather
than by crashing.

---------------

Reviewed By: rchen152

Differential Revision: D85785750

fbshipit-source-id: 9f734a10651afd8b7e0ce99e39738dc29cfd19b9