Pin to the prior type when loop recursive solves against another Var
Summary:
In order to fix the "inconsistent types breaking cycles" bugs we still have
on loops with reassignments, I want to make a change where a `Type::Var` for a
loop recursive variable behaves more like a concrete type.

Eventually, that concrete type can accumulate assumptions in the form of either
a pin or upper bounds to be checked later, but a first baby step that is needed
to unblock other changes is that we don't want these kinds of variables to
unify in the union-find algorithm.

If we treat a loop recursive var as an unknown type with a guess, and we handle
type-dependent logic in the loop (like assigning `x = f(x)` in a function) with
a "single-shot, optimistic analysis" where we assume the analysis using the
type from above the loop will work, then:
- for a case like `x = f(x)` we can just accumulate an upper bound
- but for other cases like method calls, e.g. `x = x.do_something()`, we
  really have to constrain it to be the exact type of our best guess - no other
  option really exists.

That leaves open the question of how to handle something like `x = id(x)` where
`def id[T](_: T) -> T` where a loop recursive var passes through a generic function.

In principle, we could handle this by propagating the loop recursive var and
accumulating arbitrary bounds (upper and lower - we can get lower bounds if we
try to do this because in a more complex function call the loop recursive var
can get unified with a type var used in multiple checks that are both covariant
and contravariant) at which point we would basically be doing
whole-function-constraint-solving-style analysis for any variable assigned in a
loop.

But that would be complicated, and since loops can be pretty big it effectively
means committing to nonlocal type checking for loop-recursive vars, with all
the action-at-a-distance implications that led us to choose bidirectional type
checking for Pyrefly.

If we don't want to do that, the obvious alternative is the one taken here: treat
Oany use of a loop recursive var inside the solve of a type variable, e.g. in
`x = f(x, y)`, as a "read" of the type of `x` just like we would for a method
access. That means we should just pin the type.

A big consequence of this will be that `LoopRecursive` vars cannot wind up
in upper-bound positions which will greatly simplify the implementation of the
simple behaviors I am hoping to get because we don't need both upper and lower
bounds or the ability to unify two loop recursive vars all we need are upper bounds.

Reviewed By: rchen152

Differential Revision: D85785753

fbshipit-source-id: aa8f4d895a05a2629725633009d5cc08fb657e71