Create DynamicFallbackSearchPath
Summary:
This diff creates a variant and inner struct for `DynamicFallbackSearchPath`s, which are fallback search paths that can be created for any file from its parent directory up to a specific root or filesystem root.

This will be used for constructing and caching ancestors of a directory, which we need and use in `standard_config_finder`, but will also be useful in build system configs for producing filesystem fallbacks for non-build system files in LSP mode.

Reviewed By: kinto0

Differential Revision: D85473674

fbshipit-source-id: 40f4d4321453fe78446a77616e5c36b62a0bddf6