Add a test of Airflow-like use of bitshift
Summary:
A request from Airflow is that we make sure we avoid erroring on the use
of `x >> y` as an expression whose value is discarded; they use this
syntax to express DAG dependencies, so the expression is side-effecting
and a number of linters complain.

I'm not entirely sure we need this test since I don't think Pyrefly would
ever implement an unused expression lint, but I figured I'd put it up
for review; it might be nice to be able to point to it in discussion with
partners.

Reviewed By: ndmitchell

Differential Revision: D86310205

fbshipit-source-id: 692bc7dbaba613e0fd6f99887e2329d54c12b60d