preserve notebook contents in didChange
Summary:
There was a bug lower in this diff stack where didChange would actually set all cell contents to empty string, except for any cells that were added in the change request.

This diff preserves existing cell contents, and also refactors the code to avoid unnecessarily turning a map into a vec and back to a map.

Reviewed By: kinto0

Differential Revision: D86064878

fbshipit-source-id: 0cf742a10490e9d87b65afd912d5a2b3ef17ff5f