Fix adding new suppressions
Summary:
Right now, our suppression tool will add additional comments instead of editing existing ones. This diff aims to make the suppression logic a bit more comprehensive, and will add new codes to lines when two errors are found on the same line.

right now when a new error is introduced, we add an entirely new line:
```
# pyrefly: ignore [bad-return]
# pyrefly: ignore [missing-attribute]
```

this ensures we're are editing existing suppressions when possible:

```
# pyrefly: ignore [bad-return, missing-attribute]
```

Reviewed By: stroxler

Differential Revision: D86219177

fbshipit-source-id: 0fee33f70dcf6df28609933520a63e9450c853bd