Cache interpreter stdlib paths in environment
Summary:
In order to properly determine if a file is part of the Python stdlib, we have decided to keep a cache of stdlib paths in the Python environment. This diff sets up the cache that will be used in future diffs.

We add a new SmallSet which will be used as a cache as well as one function to insert values into the cache and one to get the results of the cache. The insert function is called at the end of get_env_from_interpreter. The get function currently unused but will be called in later diffs.

This result came after various discussions on diffs/posts as well as in person

- https://www.internalfb.com/diffs/browse?start=0&phabricator_diff_id=3410242412465971
- https://www.internalfb.com/diffs/browse?phabricator_diff_id=584297344740102
- https://fb.workplace.com/groups/657557118131740/permalink/1870470923507014/

Reviewed By: connernilsen

Differential Revision: D86334308

fbshipit-source-id: aa5f0a050fc6401a8d0b54300d29df4dff54ac41