Fix materialization of *args and **kwargs with type Any
Summary:
Fixes https://github.com/facebook/pyrefly/issues/1487.

Fixes a bug involving a Materialization type showing up where we expect an iterable (for *args) or a mapping (for **kwargs).

This fix (replacing a top-level Any with the expected Iterable/Mapping before calling materialize()) is a little clunky, but from a practicality pov, I think this is the best approach, because it limits where in the codebase Type::Materialization can show up and therefore where we need to support it.

Reviewed By: stroxler

Differential Revision: D86443699

fbshipit-source-id: dbc438945c36736fefed2f3cf510441deae7adbc