Implement the solver semantics for LoopRecursive
Summary:
This diff implements the core semantic analysis change proposed in
https://docs.google.com/document/d/1QRY6E8Bhy7a0atVio-YnzJlaqa3jFcM8DwOUV6HmwFE

We treat `LoopRecursive` variables, which model the types of variables that
are reassigned inside a loop, differently:
- When they are forced, we pin them to the prior type
- When they are used in a generic function call and have to be solved against
  a type variable, we pin them to the prior type
- When they are used in any other context, we check whether the use (e.g.
  calling in a non-generic function or assigning to an annotated variable),
  we check whether the use is legal with the prior type, and if so we record
  a *bound* for later verification.

The bound kind behaves idempotently in that:
- the order of `bounds` should not affect the final errors produced
- if we ever pin the type to the loop prior, all bounds are ignored
  (because the pin to the loop prior is always more restrictive than
  any bounds).

When the cycle involving a `LoopRecursive` finishes, we validate the bounds
(or the pin) against the final answer. For now, we still use the same
"inconsistent types breaking cycles" error as before - this diff is
implementing only the sematnic improvements we want, not the UX improvements
which will require tracking more text ranges and implementing new error messages

Fixes https://github.com/facebook/pyrefly/issues/726

Note that it does *not* fix the bug in 747; this is because we really need to
treat narrowing similarly to attribute assignment, it's a sufficiently
type-dependent operation that it needs to force Var, otherwise we wind up
with narrowing operations (silently) treating `Var` as not having the semantics
that the eventual type would have, which is confusing and quite difficult to
debug unless you already know what's happening. I'll make that change in the
next diff.

-----------

Reviewed By: rchen152

Differential Revision: D85922045

fbshipit-source-id: 96e3c3843141c6d558483b1ee816b340de4b53b2