Add a test of `LoopRecursive` interacting with `Quantified`
Summary:
In a side conversation, samwgoldman brought up the questin of what happens
when a `LoopRecursive(prior_type, bound)` is matched against a type that
contains a quantified, like a `list[T]` in a generic function call.

My immediate answer was that this case never directly came up, so the
behavior must be emergent from how we defined primitive behaviors and that
I should check.

This test verifies the behavior, which after some thought is not too hard
to explain:
- The outer types don't involve `Quantified`, so we're going to enter the
  normal logic for comparing a `LoopRecursive` to any concrete type: we
  perform an `is_subset_eq` of the prior type against the type being matched
  (here, a `list[T]`) where I use the `T` to emphasize that this is a
  `Var` created for `T` in this one generic call. That subset check very likely
  will solve the `T` - in this case `T` gets solved to `int`.
- Assuming the subset check passes, we will also record a constraint that
  the loop recursive var (so the types of `good` and `bad` at the top of the loop
  here) is less than `list[T]`. Since the solver produces an answer for `T`
  (in this case, it gets solved to `int`), in this case that means we accumulate
  a bound of `list[int]`.

As a result, similar to other examples we get the behavior where as long as the
final `Phi` winds up being compatible with the bound produced in the generic
function, the code will type check but any reassignment flowing into the Phi is
incompatible we'll get type errors.

-----------

**A digression: the bindings graph can be tricky to reason about**

One thing that was surprising to me at first is that we're type checking the line
`bad = f(bad)` twice; this is visible from the test if you know what's going on
because:
- we know `bad = f(bad)` actually *passes*, because that's the only way
  we would have known to pin `T` to `int`
- but we also get a type error where Pyrefly knows that `bad: list[int] | list[str]`,
  which is only possible to know *after* we've solved the loop phi

How can this be? It's a consequence of the way `Anywhere` works: every
assignment currently flows into the `Anywhere` node for `bad`, which means we
jump straight into the middle of the loop. Since the cycle gets broken at
the loop Phi, that means we type check `bad = f(bad)` once inside of the loop
and then again after the cycle is completed. The second check produces different
type errors.

This behavior is very close to some of the nondeterminism behaviors we can
observe due to `Var` (it's very possible to get exactly the same sort of thing
nondeterministically through forcing an export), although in this case it's
deterministic and happens even single-threaded because of the shape of the
binding graph.

I'm not sure whether we actually want to change this behavior. The existing
semantics, which try to be pretty close to sound, require making sure that all
assignments flow into the Anywhere phi; in theory we could try to prevent
assignments that occur inside of branching control flow from flowing *directly*
into Anywhere (instead, we could make them flow in via the merge), but the
problem is that it isn't easy to know at the point where we see such an
assignment whether or not it will wind up in the merge, because straight-line
control flow might reassign it.

Reviewed By: rchen152

Differential Revision: D86233757

fbshipit-source-id: 1cfa23c15267d6b59635edad25c7b89bd3badb48