Force types flowing into narrows (including boolop)
Summary:
The previous diff put the plumbing in place needed to fix all currently
open bugs on valid loops that reassign variables (there are type-dependent
cases stemming from generic calls and overloads no one has complained about
that it doesn't fix).

But there's a problem with narrowing that was preventing me from closing
issue #747: at the moment, an unforced `Var` never narrows, which means
both narrowing itself and the "narrow" operations applied to boolean
expressions in non-tail positions of expressions like `x or 1` don't
actually occur.

This is clearly not something we can allow. My solution is to treat narrowing
similarly to attribute resolution: just treat a narrow against a type (or its
use as the prefix of a bool op, which requires a narrow of the result) as
forcing the type.

This fixes the behavior reported in #747 for loops. One could argue that we
shouldn't force other kinds of `Var`s but the behavior if we did that would
be unpredictable (for example a recursive type would fail to narrow, but might
then later get populated by `int | None`, which is very confusing and impossible
to debug unless the person debugging understands the flaws of Var pinning very
well). I think this approach is better; I discussed with rchen152 and she also
thought forcing all `Var`s is probably preferable.

Fixes https://github.com/facebook/pyrefly/issues/747

Reviewed By: rchen152

Differential Revision: D85922046

fbshipit-source-id: 7dd46a89572751da843f7d174e132da57ef42178