Emit error when writing to a final field in a dataclass (#1509)
Summary:
When a final field of a dataclass is not initialised, the field is not marked as `Readonly` but as `Readwrite`. In these cases, we lookup if the field being written to is a dataclass field, and emit an error if the field is final.

We also handle the case we modify a final field in dataclass of the class instead of an instance.

To do so, we need to pass on the class base in a helper function, alongside the instance class. Maybe we should use an Either type since the two should not be passed at the same time.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1509

Reviewed By: grievejia

Differential Revision: D86487587

Pulled By: yangdanny97

fbshipit-source-id: f5dfedee9758d183df4a10cb64a839e1d7d26488