Add support for built in types and third party typeshed types on hover.
Summary:
This adds support to hover for built in types. When you hover over a type that is built in, you will see those types as go to type suggestions.

Will also provide suggestions for any types that are bundled with third party typeshed. Note that these third party types will only be used when a config file is not present in the project.

We also refactor out the to_real_path function to be used in both a wasm and non wasm context.

{F1983344942}

{F1983344943}

{F1983345112}

Reviewed By: kinto0

Differential Revision: D86524079

fbshipit-source-id: cf717aff7dd16bff7949b088143ba8099c16c205