Implement goto type definition with collect_symbol_def_paths
Summary:
Although there are a lot of lines in this diff the majority are just tests so it should not be too difficult to review.

Here we re-use the shared collect_symbol_def_paths in order to power goto_type_definition in the IDE. In the goto_type_definition function, we first look for the type at the particular handle. If we find a type, we pass it into collect_symbol_def_paths and collection any locations. If the collection we get back is non_empty we will convert these into a collection of TextRangeWithModule and return early. Otherwise, we fall back to the existing find_definition functionality.

Reviewed By: kinto0

Differential Revision: D86532031

fbshipit-source-id: aac1d1602f1c0503c9376cc6e35d9b28e3ade8a3