Check for duplicate keys when collecting key-values into a map
Summary:
# Problem
We often use `.collect()` to turn an iterator of (key, value) pairs into a hash map.
However, the iterator might have multiple pairs with the same key, which would most likely be a bug in our logic.

# Solution
This diff implements `.collect_no_duplicate_keys()` that is similar to `.collect()` but fails if multiple pairs have the same key.

Reviewed By: tianhan0

Differential Revision: D86599290

fbshipit-source-id: 38c91c1feaf7cb4513b1cb16c6dd4f39301b5954