Remove some redundant checks
Summary:
clean_kirby

With D86427192, we can see from function names that within `local_references_from_local_definition`, the module path of the current file must be the same as the module path of the definition (because the def is ensured to be local). Therefore, any further check on whether those two are the same is redundant and can be safely removed.

Reviewed By: kinto0

Differential Revision: D86523649

fbshipit-source-id: 76647f504bf68d6250a0c05a59dedf579c12a57c