Factor `local_references_from_definition()` into 2 smaller functions
Summary:
justmovingthingsaround

There's a block of code in `local_references_from_definition()` which handles find reference for a definition defined outside of the current module, and there's also a block of code there that handles find reference for a definition defined within the current module. Both blocks are somewhat sizable and hence it becomes cleaner to factor them out which serves as an explicitly document on what's going on there. I've also renamed some callees on the find-ref-for-local-def side to make it more clear that those logic only applies for local defs.

Another nice benefit is that we could extract sort-and-dedup logic out and avoid duplicating it.

Reviewed By: rchen152

Differential Revision: D86427192

fbshipit-source-id: b017e4db504b940880b8a90ce1acb80832c643b4