Filter project excludes for site package path entries if they would exclude search path
Summary:
Pyrefly has some logic to add all `site-package-path` entries to `project-excludes`, but sometimes the Python interpreter includes the root of the project. Historically, we've filtered this out by using the project root, but if the actual project root differs from what Pyrefly thinks it is, we might end up excluding the project by default.

This happens in legacy projects, where instead of using `src` as the project directory, `lib` is used.

This diff instead uses all `search-path` entries, not just Pyrefly's heuristic project root, to determine if a site package path should be added. This way, anything that might be considered "project code" because it's in `search-path` will always be able to be included in `project-includes`.

Fixes https://github.com/facebook/pyrefly/issues/1498

Reviewed By: kinto0

Differential Revision: D86682740

fbshipit-source-id: c232c0fab3b9616f9884a2c7784795349e2b8ce7