Replace Variable::Unsolved with Never when it appears in a union in a return type
Summary:
This is the first of two diffs that aim to fix pyrefly's behavior in this scenario:
```
def f[T](x: int, y : T | None = None) -> int | T: ...
reveal_type(f(0))  # what should this type be?
```
Mypy and pyright both say that the revealed type is `int`, but pyrefly says `int | Unknown`. This difference in behavior is causing issues like https://github.com/facebook/pyrefly/issues/1395 and https://github.com/facebook/pyrefly/issues/1114 as well as a failure in the django-stubs typing tests.

As far as I can tell, mypy always falls back to `Never` for unsolved type parameters, while pyright falls back to either `Never` or `Any` depending on whether the unsolved parameter is in a union. An attempt I made to match mypy's behavior introduced a ton of spurious type errors, so this diff tries to match pyright's behavior instead. I did this by replacing the `expand_vars` call we do on the type returned from a function call with a call to a new `finish_function_return` function that does var expansion and also replaces unsolved parameters in unions with `Never`.

Reviewed By: stroxler

Differential Revision: D86014506

fbshipit-source-id: 5aa8fcbeb4a1ebabe7e7c3abfa63d065b1dc2f74