Add test that asserts that renaming third party symbols from venv packages currently works.
Summary:
This is the first diff in a stack to address the github issue at https://github.com/facebook/pyrefly/issues/1433

This issue asserts that you are currently able to rename package from third party package. Other language servers like Pylance do not allow this and will show an error message if this is attempted.

This diff just creates a test which covers the existing behavior and shows that renaming third party symbols is currently allowed. Subsequent diffs will instead show an error message and the tests will be updated to reflect this new behavior.

Reviewed By: kinto0

Differential Revision: D86768501

fbshipit-source-id: 072bf1244c438cee7f95c6e8727839785f5e7708