Add flag to disable excludes heuristics
Summary:
# This diff...

changes:
1. Adds a new flag and config option `disable-project-excludes-heuristics` (definitely open to changing the name.
2. uses the flag to figure out when we shouldn't use the `ConfigFile::required_project_excludes()`
3. uses the flag to disable automatically appending most site package path values to `project-excludes`

# Background...

Several reports came in mentioning that type checking files under dotted directories broke type checking for them. This happens because the new, always-included `project-excludes` exclude those directories.

There are several approaches we can take here:
- undo hard-coding some project excludes
- addd a flag to disable project excludes heuristics
- add a new flag to overwrite project excludes with
- analyze the included globs and use that to determine when we should ignore project excludes

After some discussion on one of [the issues](https://github.com/facebook/pyrefly/issues/1519), I've come to the conclusion that we should probably just make this a flag, which I mentioned as a possibility of something we'll need on the original diff (D85266724, 60ace40). I'm also open to undoing the hard coding of it, but I think it's generally a good thing to have, since users probably don't want to rewrite that part themselves when overriding excludes (it's very likely to be more common than needing to type check stuff in excluded directories).

I'm definitely open to discussing the other alternatives, but I think this approach is probably the best.

Reviewed By: kinto0

Differential Revision: D86692262

fbshipit-source-id: b4c1644181ead617e5e8f3365897421840c7537a