Remove `Name` field from definition metadata
Summary: I noticed that `DefinitionMetadata` sometimes retains a copy of the variable name. This is not necessary given that (1) we've already have the range of that name, and (2) in all places where the variable name is required, we also gets access to the module info which contains the raw text of the file. That means variable names can always be obtained from range + raw text, which renders the info redundant on `DefinitionMetadata`

Reviewed By: stroxler

Differential Revision: D86719936

fbshipit-source-id: 7f61bc468e6bacc5af95853d84c71e03449e8e40