Do not allow renaming of third party symbols.
Summary:
Now in the prepare_rename lsp function, we first ensure that the symbol we are trying to rename is not a third party symbol. To do this we first gather all the definitions. If any of these definitions belong to a third party module, we return None.

This also fixes the issue where when w

Right now this diff means this operation will just fail silently. In a subsequent diff we will make sure that this provides an error message as well.

 {F1983433472}

Reviewed By: kinto0

Differential Revision: D86784199

fbshipit-source-id: 00849d2a99f5dce73b53b03d739c06dd6eb76e4c