Modify typeshed upgrade script to take on `docify` dependency
Summary:
In order to resolve https://github.com/facebook/pyrefly/pull/1446 we need to run `docify` on typeshed. The best time to run it would be at typeshed download time, since that would avoid the cost at Pyrefly startup.

Given that `docify` is a 3p library, our typeshed update script can no longer be self-contained. Fortunately, `uv` supports inline script metadata and we could rely on that to simplify dependency management when running the script in OSS mode.

This diff adds the dependency metadata to our upgrade typeshed script, and update our docs to provide details on how to run the script now that it's no longer self-contained. No behavior changes yet -- the new dependency will be utilized in the next diff.

Reviewed By: kinto0

Differential Revision: D86799251

fbshipit-source-id: 163026df9f9976f217f7217078ce689741365e8c