Run `docify` on typeshed stubs after downloading
Summary:
Following up on D86799251, we add a new flag to `update.py` indicating whether to docify the downloaded typeshed (default is true).

It takes a while to docify everything so I think it's worth adding a flag that turns the behavior off. Also, I'm only docifying `stdlib` for now. It's possible to docify 3p libraries as well but that requires the update script itself taking in all those 3p dependencies which, if we want those in the internal Buck build as well, would take much more work. Let's start with stdlib first.

Reviewed By: kinto0

Differential Revision: D86799254

fbshipit-source-id: 1ebd0e3d749477506e83287eb61481d1885ae2b5