Add test to show that renaming named function argument only updates the definition.
Summary:
This test covers the second comment on the github issue found here: https://github.com/facebook/pyrefly/issues/1433

I think the comment on the github issue is entirely clear as to what is happening, but what I have seen is that trying to update a kwarg to a function will only update the name in the function definition and not where the function is actually invoked. This diff adds a test to cover this logic and will be updated once a fix for this is implemented.

{F1983434517}

Reviewed By: kinto0

Differential Revision: D86795066

fbshipit-source-id: 56b7a31a24533a7846528f31d632bb84097cf43a