Ignore __pycache__files in didChangeWatchedFiles events
Summary:
This diff tries to implement the non-globbing parts of D85953911 (which I don't think can be done). We instead try to handle incoming requests to not process any changes in `__pycache__`directories.

I'm not sure if we should also add the early exit functionality to each LSP function, since it seems to me it would be rare for someone to open something in `__pycache__`, and if they do, performing some operations on it wouldn't be too resource intensive. Especially since the number of opened `__pycache__` files would likely be orders of magnitude smaller than the number of those being updated in a large project by Python itself.

Reviewed By: kinto0

Differential Revision: D86828413

fbshipit-source-id: 488f12a07234040d199571453046753f61884a22