go-to definition / hover on methods of .pyi
Summary:
D80145015 implemented .py preference on go-to definition, but it did not work for attributes. that's because when we solve a type `a`, `a.attribute` doesn't do another find, it simply goes to the definition of a from solving which is always the .pyi.

asuka implemented a fix for that [here](https://github.com/facebook/pyrefly/pull/1520) for hover docstrings. the intution is to do a second find on attribute definition lookup and look at the AST to find the corresponding definition. this works pretty well, and this diff stack generalizes it to definition.

Reviewed By: connernilsen

Differential Revision: D86814372

fbshipit-source-id: 8b7ed56d74071db55188d448f1eb76f91bcc612d