Check for disjoint bases in `intersect_impl`
Summary:
Teach pyrefly that intersecting disjoint bases produces `Never`. As a standalone change, this isn't all that useful, but if we add a true intersection type, we'll need this logic to avoid producing impossible intersections.

For now, I've added a very basic check that doesn't take inheriting from a disjoint base into account, and left a TODO to implement the full disjoint base calculation from PEP 800.

Reviewed By: grievejia

Differential Revision: D86880954

fbshipit-source-id: 652078bd004ee7e7cc5c67988d12ecd2a964a2a8