Add function to collect keyword arguments based on name.
Summary: This diff introduces a function which when given a name will find all local uses of that name which are keyword arguments. In a future diff, this function will be used to collect all those references, which will will then be narrowed down to only those which belong to the defining function.

Reviewed By: kinto0

Differential Revision: D86868371

fbshipit-source-id: dfeedd1ab85f0818916343f6bbd1b92b21192dc9