Ensure that renaming a keyword argument also updates the names in function calls within the same file.
Summary:
This builds upon the work in previous diffs and implements the local_keyword_argument_referneces_from_parameter_definition in local_references_from_local_definition. In the variable and VariableOrAttribute match arms, we do another check do see if the SymbolKind is either a parameter, variable, or None. If any of these cases are true, we will also include the keyword arguments in the collection that we return.

{F1983467395}

Reviewed By: kinto0

Differential Revision: D86888748

fbshipit-source-id: 2d51165b7d6df1afab8e3c360af332e65ecedd49