Err when a scoped type parameter shadows an enclosing scope.
Summary:
It makes perfect semantic sense to allow shadowing with scoped type parameters,
but (I suspect partly because of usability concerns coming from folks used to
the legacy syntax) the spec says type checkers must complain here.

Implement the check. We use a direct scan of the enclosing scopes rather than
looking at the currently bound value because (a) this lets us do it at bindings
time instead of solving time, and (b) it avoids a false negative if the name is
also shadowed by some non-type-parameter value in an intervening scope (which
is legal, and the conformance tests explicitly require us to complain in this
case).

Reviewed By: rchen152

Differential Revision: D86979784

fbshipit-source-id: c8d2e7a551864e6094e971b32a288c9ed59574bc